requirements.txt
streamlit>=1.30.0
llama-cpp-python>=0.2.64
